+++ packages/uniapp-components/checkbox-group/type.ts
@@ -4,6 +4,7 @@
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
 
+import type { CheckboxIconType } from '../checkbox/type';
 import type { KeysType } from '../common/common';
 
 export interface TdCheckboxGroupProps<T = CheckboxGroupValue> {
@@ -66,6 +67,7 @@ export interface CheckboxOptionObj {
   label?: string;
   value?: string | number;
   disabled?: boolean;
+  icon?: CheckboxIconType;
   checkAll?: true;
 }